Quick and dirty hack. Stop testing for mouse locations if they were already found.
git-svn-id: https://svn.aros.org/svn/aros/trunk@28728 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/necessary/USB/classes/HID/hid.h
@@ -125,7 +125,7 @@ typedef struct MouseData {
     char                        *report;
     uint16_t                    reportLength;
     uint16_t                    nreport;
-    uint16_t                    mouse_report;
+    int                         mouse_report;
     
     struct Process              *mouse_task;
